improves tests
